+++ config/config.go
@@ -8,11 +8,12 @@ import (
 )
 
 type Config struct {
-	GIN_MODE    string
-	APIURL      string
-	APIKey      string
-	DBURL       string
-	NEON_DB_URL string
+	GIN_MODE      string
+	APIURL        string
+	APIKey        string
+	DBURL         string
+	NEON_DB_URL   string
+	APIKey_GOOGLE string
 }
 
 func LoadConfig() Config {
@@ -26,10 +27,11 @@ func LoadConfig() Config {
 
 	// fmt.Println("🚀 Đang tải cấu hình từ biến môi trường...", os.Getenv("NEON_DB_URL"))
 	return Config{
-		GIN_MODE:    os.Getenv("GIN_MODE"),
-		APIURL:      os.Getenv("API_URL"),
-		APIKey:      os.Getenv("API_KEY"),
-		DBURL:       os.Getenv("DB_URL"),
-		NEON_DB_URL: os.Getenv("NEON_DB_URL"),
+		GIN_MODE:      os.Getenv("GIN_MODE"),
+		APIURL:        os.Getenv("API_URL"),
+		APIKey:        os.Getenv("API_KEY"),
+		DBURL:         os.Getenv("DB_URL"),
+		NEON_DB_URL:   os.Getenv("NEON_DB_URL"),
+		APIKey_GOOGLE: os.Getenv("API_KEY_GOOGLE"),
 	}
 }