+++ internal/controller/farm.controller.go
@@ -2,6 +2,8 @@ package controller
 
 import (
 	"avatar/internal/service"
+	"encoding/json"
+	"fmt"
 	"net/http"
 
 	"github.com/gin-gonic/gin"
@@ -15,4 +17,23 @@ func GetFarmData(c *gin.Context) {
 	}
 
 	c.JSON(http.StatusOK, farmData)
+}
+
+type SyncAnimalRequest struct {
+	ID         int    `json:"id" binding:"required"`
+	Name       string `json:"name" binding:"required"`
+	Desc       string `json:"desc"`
+	BornTime   int    `json:"born_time"`
+	MatureTime int    `json:"mature_time"`
+}
+
+func SyncAnimal(c *gin.Context) {
+	var param SyncAnimalRequest
+	if err := c.ShouldBindJSON(&param); err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid request body"})
+		return
+	}
+	jsonBytes, _ := json.MarshalIndent(param, "", "  ")
+	fmt.Println("Received animal JSON:\n", string(jsonBytes))
+	c.JSON(http.StatusOK, gin.H{"message": "Sync animal success"})
 }