+++ internal/router/global/global.router.go
@@ -9,4 +9,7 @@ import (
 func GlobalRoutes(r *gin.RouterGroup) {
 	r.POST("/image/b2i", controller.Byte2Image)
 	r.POST("/captcha/train", controller.TrainCaptcha)
+	r.POST("/dialog/input/answer", controller.GetAnswerInputDlg)
+	r.POST("/crawl/icon", controller.CrawlIcon)
+	r.POST("/crawl/image", controller.CrawlImage)
 }